Statement-3-Placement-Predictor/backend/requirements.txt
fastapi
uvicorn
anthropic
scikit-learn
pandas
numpy
faiss-cpu
sentence-transformers
pymupdf
python-multipart
pydantic
joblib
python-dotenv
python-docx
